+++ TDTChocolate/FoundationAdditions/NSDictionary+TDTSubsetEquality.h
@@ -3,12 +3,14 @@
 @interface NSDictionary (TDTSubsetEquality)
 
 /**
- Recursively compare the structure and values of two dictionaries to
- determine if one of them contains a subset of the entries (keys, values) of
- the other. If any of the values are an array, then the array contents are
- also recursively matched.
+ @return YES if the all of the keys in the receiver are present in @p dict,
+ and for each such key, the corresponding values recursively return YES
+ when probed using @p tdt_isSubsetOf.
+ 
+ NSDictionary and NSArray instances respond to @p tdt_isSubsetOf in a custom
+ manner. For everything else, it fall backs to an @p isEqual check.
 
- @return YES if the receiver is a subset of the argument @p dict.
+ @see `- [NSArray tdt_isSubsetOf:]`
  */
 - (BOOL)tdt_isSubsetOf:(NSDictionary *)dict;