+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Fix variable name in macro.
   * Add method to check blank string.
   * Add method for array reduction.
+  * Fix nil handling in custom asserts.
 
 2.5.1, 22-July-2014