+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-2.5.?, 22-July-2014
+2.5.?, ??-July-2014
 
   * Fix variable name in macro.
   * Add method to check blank string.
   * Add method for array reduction.
   * Add method for reversing arrays.
+  * Expose isSubsetOf for arrays.
 
 2.5.1, 22-July-2014